Task 6 made square_generator.py and imported SquareGenerator into main
